+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 - Added methods `doi` and `citation` to `DataCollection` class.
   ([#203](https://github.com/nsidc/earthaccess/issues/203))
   (@Sherwin-14, @chuckwondo)
+- Added property `__geo_interface__` to `DataGranule` class.
+  ([#1179](https://github.com/nsidc/earthaccess/pull/1179)) (@s-kganz)
 
 ### Removed